Refactor To Follow Best Practices
ChannelsFragment.kt:

In this class, instead of calling ApiHandler.getChannels(workspace.id) directly from the Fragment, I leveraged ChannelsViewModel to perform this operation. This way, the Fragment only interacts with the ViewModel and isn't directly responsible for retrieving data.

ChannelsViewModel.kt:

In the ChannelsViewModel, the name of the LiveData and methods didn't reflect that they're dealing with Channels, so updated them to make the code more understandable