Member Provider was not thread safe so it caused flaky test cases
I've introduced a private DispatchQueue (membersQueue). Anytime you access or modify the members dictionary, you do it inside of a membersQueue.async or membersQueue.sync block. This ensures that only one block of code can access or modify members at a time, even if loadAllMembers(withToken:) and memberName(forID:) are called from multiple threads at the same time.

Note: It's crucial that you update the @Published property on the main queue because SwiftUI relies on this behavior.