Fixed off by one error with my view compensating
Explanation :

generate() function was returning April for the month of 5, so to get may i had to pass in 6 which is suppose to be june, my view I compensated for that, but the logic was wrong.

This fix now makes passing in dates indexed by 1.

When JS was indexing by 0.